fix(profiles): avoid config import cycle
